Fix broken tests due to refactor

+++ internal/fixture/processor.go
@@ -55,7 +55,7 @@ func (s Processor) currentSeason(done chan bool) {
 	go s.callClient(ids, done)
 }
 
-func (s Processor) callClient(ids []int, done chan bool) {
+func (s Processor) callClient(ids []int64, done chan bool) {
 	q := []string{"fixtures"}
 
 	for _, id := range ids {
@@ -71,7 +71,7 @@ func (s Processor) callClient(ids []int, done chan bool) {
 			s.handleFixtures(res.Data.Fixtures.Data)
 
 			defer waitGroup.Done()
-		}(id)
+		}(int(id))
 	}
 
 	waitGroup.Wait()
